Remove redundant "lockfile-explorer start" subcommand, restoring the old CLI syntax
